requirements.txt
fastapi
uvicorn[standard]
x402[fastapi]
python-dotenv
